+++ src/Local_Development/Settings.php
@@ -246,6 +246,12 @@ class Settings {
 			'production'  => __( 'production', 'local-development' ),
 		];
 
+		$environment_type = defined( 'WP_ENVIRONMENT_TYPE' ) ? \WP_ENVIRONMENT_TYPE : false;
+
+		self::$options['extras']['environment_type'] = ( ( new Extras( self::$options ) )->is_localhost() && $environment_type !== self::$options['extras']['environment_type'] )
+			? $environment_type
+			: self::$options['extras']['environment_type'];
+
 		echo '<div style="float:left;margin-top:8px;">' . esc_html( $args['name'] ) . '</div>';
 		foreach ( $environments as $key => $value ) {
 			$checked = isset( self::$options[ $args['type'] ][ $args['id'] ] ) && $key === self::$options[ $args['type'] ][ $args['id'] ] ? 'checked' : '';